BBB/AM335x: Store addresses of custom exception handlers in the Public RAM reserved for exception vectors
The AM335x jumps to a predefined location when handling an exception.  The default locations of the
exception handlers are tablified in Table 26-3 under Section 26.1.3.2 in the AM335x TRM.  When Umon
is executing and an exception occurs the AM335x will normally try to jump to the default exception
handler appropriate for the exception that has occured.  Since Umon defines a set of custom exception
handlers, the addresses of these exception handlers must be stored in the RAM exception vector
space in the Public RAM of the AM335x.  This is done so that when an exception occurs, instead of
program execution being transferred to one of the default exception handlers, it will instead be
transferred to a custom exception handler defined by Umon.