BBB: Build an MLO file when booting from non-XIP devices (e.g. SD)
An MLO file will be created from the concatenation of a corresponding GP header
and the uMon image.  A GP header must be prepended to the uMon image when
booting from non-XIP devices.  Refer to the AM335x TRM, Section 26.1.9 for more
information.

These type of images have been tested when booting the MLO file from an SD card
with a FAT32 primary partition marked as Active.